Add video-derivatives bucket
To be used for HLS, but we'll name it video_derivatives generically, in case we have other derivatievs for download later -- our motivations for putting things in separate bucket would apply to such too, and it would just make sense to consoloidate them. Will name keypaths in bucket accordingly, for future expansion.

This PR does NOT set up a backup bucket with replication. Our initial discussions were that we did not need/want such. But we should defintiely go over it with Chuck/@CHF-IT-Staff

The bucket IS set to public read, like existing derivatives. We don't actually have a good way to provide signed url access to HLS, it wouldn't work as static files.