Include app s3 bucket name ENV vars in terraform output
They were one of the first things we did and were 'obvious' so we just did them manually -- but terraform output should be able to be cut-and-paste (well, with a bit of transformation) to create all app ENV that comes from terraform things, they should be included. We have too much to keep track of, the goal is as little 'manual' as possible.